Hi Lux
Everything of the best for 2015!

Please can you change the integration process for the above so that there is a check first in the MAF database whether the code_article_client already exists there.  Only if it does not yet exist, then the integration should take place.

Please call me if there is any uncertainty.