Hi Hans/Lux
We had some cases where a bin was staged on a run, say with presort_staging_child_run_id 100 on date 2015-02-13. But the bin was never tipped by MAF.
Then the same bin was staged again on a new run, say with presort_staging_child_run_id 150 on date 2015-02-18. This resulted in two records in the Apport table for that bin.

MAF then read the record where the presort_staging_child_run_id was 100, and this field, being one of the indicators of a new MAF lot, caused the creation of a new lot.

Solution:

Before integration into the Apport table takes place, the MES system should check whether the bin is already in the Apport table. If it is then check the fields that MAF completes, and if these are NULL, then the existing record should be deleted from Apport, and the new record inserted.

Example SQL statement when checking:

select * from Apport where NumPalox = '50088330' and LotMAF is null and DateLecture is null and StatusMAF is null;

So, if this returns a record, then it should be deleted, and the new record inserted.