removal of sealed_location_id check on bins before update of rmt_product(bins now to be updated irregardless of being sealed before):
Nafiesa Allie
10:59 AM (23 hours ago)
to me, Hans
Hi Netsai

With the last CR on Coldstore Management we requested that the ripe_point/rmt_product_code not be changed if the sealed_ca_location_id is not null.
Apparently, the same bin can be repeatedly sealed in different coldstores, so this check must come out.
Please can you make the necessary change?