Hi Judy
I see that when I was testing I removed the line of code that sets the orchard_code on the bins and I never put it back in. Find the attached package and deploy it.

I'll have to manually fix the 9577 records. I've also attached a bins.csv that show the recalculated bins.orchard_code values. See if these values make sense, if they do, let me know and I'll run my script to commit them into your database.