Hans,
Can you please check and see how we can enforce the cycle of a cold room further.

I know we a sequence in which the transaction statuses need to follow, but i do not think we check and make sure the status can not be followed by itself. Please see examples below.